+++ public/style.css
@@ -1,5 +1,6 @@
 body {
   font-family: "Arvo", serif;
+  background-color: #7DAD71
 }
 
 a {
@@ -130,7 +131,7 @@ img.poke {
 
 .background {
   background: linear-gradient(rgba(20, 20, 20, 0.5), rgb(255, 255, 255)),
-    url("https://i.pinimg.com/originals/06/5e/5e/065e5e9239f7d492dc3d64e57e84d6ed.jpg");
+    url("http://i67.tinypic.com/2vd1ndg.png");
   background-attachment: fixed;
   background-size: cover;
 }
@@ -149,17 +150,20 @@ img.poke {
 }
 
 .header {
-  font-style: italic;
-  display: inline;
-  padding: 15px;
+  background-image: url("background.jpg");
+  /* background-size: cover; */
+  background-position: 0 63%;
+  margin-bottom: 50px;
+  height: 250px;
+  border-bottom: 1px solid black;
 }
 
 .page-header {
   padding: 20px;
 
   background: linear-gradient(rgba(20, 20, 20, 0.5), rgba(20, 20, 20, 0.5)),
-    url("https://i.pinimg.com/originals/06/5e/5e/065e5e9239f7d492dc3d64e57e84d6ed.jpg");
-  background-size: cover;
+    url("http://i67.tinypic.com/2vd1ndg.png");
+  background-size: initial;
   background-color: #f2f295;
 
   width: 100%;
